requirements.txt
# Core dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy[asyncio]>=2.0.25
aiosqlite>=0.19.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-telegram-bot>=20.7
boto3>=1.34.0
strands-agents>=0.1.0
strands-agents-tools>=0.1.0
alembic>=1.13.0
pyyaml>=6.0.0

# Development dependencies
pytest>=7.4.0
pytest-asyncio>=0.23.0
hypothesis>=6.92.0
pytest-cov>=4.1.0
ruff>=0.1.0
mypy>=1.8.0
